content用のSQLファイルを作成した

+++ src/main/resources/db/migration/V1_2__Product.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS `product` (
     `id` BIGINT AUTO_INCREMENT,
     `name` text NOT NULL,
-    `created_by` INT NOT NULL,
-    `updated_by` INT NOT NULL,
+    `created_by` BIGINT NOT NULL,
+    `updated_by` BIGINT NOT NULL,
     `created_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     `updated_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     PRIMARY KEY (`id`)
